Incorporate COAMPS-TC wind validation with SFMR.
